company_descriptions/requirements.txt
metaflow
transformers[torch]
boto3
pandas
toolz
pyarrow
fsspec
s3fs
datasets
pyyaml
scikit-learn
scipy
wandb